Slice 6: int64 output widths through the buffer, segments, marshalling
Widens the tree driver's output-buffer counting family from int to
int64_t so the segment/buffer marshaller and its mirrored globals can
represent per-tree halo counts beyond INT_MAX, while every narrowing
back to int (HaloAux sync, the MaxFoFWorkspace estimate, and the two
on-disk writes that are permanently/temporarily 4-byte int) goes
through one shared checked helper that FATALs on overflow instead of
truncating. No on-disk or HDF5-visible change in this slice: the
binary header stays a 4-byte int and the HDF5 attribute stays
H5T_NATIVE_INT, both behind the pre-existing INT_MAX counter guard,
which keeps its exact semantics.

Changed files:
- src/core/output_buffer.h: OutputBuffer.count/.capacity and
  OutputBufferSegment.output_first/.output_count widen to int64_t;
  workspace_start/workspace_count stay int (FoF-local).
- src/core/output_buffer.c: growth arithmetic (capacity, the
  MAX_HALO_ARRAY_SIZE cap) moves to int64_t; adds
  narrow_int64_to_int_checked(), the single checked-narrowing helper
  used by every site below.
- src/include/proto.h: declares narrow_int64_to_int_checked().
- src/include/globals.h, src/core/allvars.c: NumProcessedHalos,
  MaxProcessedHalos, and TotHalosPerSnap[] widen to int64_t in both
  declaration and definition; MaxFoFWorkspace and InputHalosPerSnap
  are untouched.
- src/core/build_model.c: the HaloAux[...].FirstHalo/.NHalos sync
  after marshalling narrows the widened segment output_first/
  output_count through the checked helper.
- src/io/tree/interface.c: MaxProcessedHalos sizing arithmetic moves
  to int64_t (value-preserving for all reachable inputs); the
  MaxFoFWorkspace estimate derived from it narrows through the
  checked helper.
- src/io/output/binary.c: save_halos's loop index widens to int64_t
  (direct NumProcessedHalos consumer, with its one FATAL_ERROR format
  updated to PRId64); finalize_halo_file narrows TotHalosPerSnap
  through the checked helper into a local int before the 4-byte
  fwrite.
- src/io/output/hdf5.c: save_halos_hdf5's loop index widens to
  int64_t; write_hdf5_attrs narrows TotHalosPerSnap through the
  checked helper into a local int before the H5T_NATIVE_INT
  H5Awrite.
- tests/unit/test_output_buffer.c: adds three tests for the checked
  helper — in-range values (including both INT_MIN/INT_MAX
  boundaries) pass through unchanged, and one value past each
  boundary FATALs (via fork/exec) — proving the guard is neither
  tautologically fatal nor off-by-one.

Validation: default-pair three-tier suite green (unit/integration/
scientific); fixture-pair (halos-only/micro-uchuu-snapshot) unit
suite green; bitwise tree-path check byte-identical; HDF5 metadata
identical except the five always-excluded provenance attributes;
zero compiler warnings under -Wall -Wextra -Wshadow
-Wformat-security -Wundef; differential lint verdict: pass after two
in-surface fixes (clang-format, freopen() return-value casts) to the
new tests. Full evidence in validation.md (gitignored, PM run
artifact).

+++ src/core/allvars.c
@@ -46,12 +46,12 @@ int HDF5_n_props;
 hid_t HDF5_current_file_id = -1; /* -1 means no file currently open */
 #endif
 
-int MaxProcessedHalos;
+int64_t MaxProcessedHalos;
 int MaxFoFWorkspace;
-int Ntrees;            /*  number of trees in current file  */
-int NumProcessedHalos; /*  Total number of halos stored for current tree  */
+int Ntrees;                /*  number of trees in current file  */
+int64_t NumProcessedHalos; /*  Total number of halos stored for current tree  */
 
-int TotHalosPerSnap[ABSOLUTEMAXSNAPS];
+int64_t TotHalosPerSnap[ABSOLUTEMAXSNAPS];
 int *InputHalosPerSnap[ABSOLUTEMAXSNAPS];
 
 int *InputTreeNHalos;
